move ups_token to models
